Test for run_game
